Implement calculation of Jacobian ratio for a Quad Element
What the Jacobian method is supposed to do is calculating the RATIO between the maximum and minimum jacobian determinant of the corner points of an element. This commit fixes the method that previously only averaged the jacobian determinants of the element.

The commit also renames methods, properties, and variables to further emphasise that we are trying to find the Jacobian "ratio", not just the determinants.

Todo: Research and implement a color scheme for classifying a "good" Jacobian ratio.