WIP: Implement Jacobian for arbitrary plane elements.
Issues in CreateMesh2D makes it hard to verify results and continue current development. Since last commit both jacobian and surfaceNormal method has changed. Jacobian first transforms input to be defined in xy-plane using the CalculateLocalCoordinatesOfPlaneElement() which is an expansion of the previous CalcSurfaceNormal method.

The biggest difference is using more built in Rhino methods for working on geometry (and transformations specifically).