Change handling of negative values of Jacobian
Instead of taking the absolute value of each determinant and displaying a remark, we now keep the sign of the determinant.

This is done so we are able to get both a valid jacobian ratio for elements where each determinant is flipped, in addition to an invalid jacobian ratio for elements that are grossly disformed and has determinants that are both positive and negative.