:bug: Make CullPointsOutsideSurface check if points are on the edge
Create a HashSet in the CullPointsOutsideSurface containing the edge points. In addition to checking if points are on surface (as before), we also make sure that the point is not equal to one of the edge points.

The reason we had not discovered the bug before likely came down to the fact that the GetBoundingBox function inside CreateBoundingBoxFromBrep had the "false" accurate tag. This creates an almost perfect bounding box for a Brep that is created from a set of points.

However, when we create a surface directly in Rhino, for some reason this _does_ create a perfect bounding box, which means that in cases where a corner is the extreme in both the x- and y-directions, a point would be created in this corner _both_ in the CreateEdgePoints and CreatePointGrid-funtions. As our code did not account for this (and the author did not create surfaces in this manner while writing the methods), the bug was not discovered until now.