Interpolate more than one point
Using a quite naive approach; after a point has been interpolated
we remove the current minimum grid point and all its' neighbours.
Because of the size of the grid, this should not cause any issues when
interpolating successive points.

Problems could arise around the boundary of the contour, but a good
mesh should not have too many points there anyway (simplified).